BUG: missing cast
- fixed nasty bug: forgot to cast to unsigned int in the call
    Comm::write((unsigned int)message.size()+1)
  Consequence was that Logger::log_master(...) only printed
  empty strings

+++ kernel/logger.hpp
@@ -327,7 +327,7 @@ namespace FEAST
 
       // write length of the log message to the buffer (add 1 to the length since string::c_str() adds null termination
       // symbol)
-      Comm::write(message.size()+1);
+      Comm::write((unsigned int)message.size()+1);
 
       // write string itself to the buffer
       Comm::write(message);
